+++ src/constants.ts
@@ -1,13 +1,3 @@
-export enum AFConstants {
-  onConversionDataSuccess = 'onConversionDataSuccess',
-  onConversionDataFail = 'onConversionDataFail',
-  onAppOpenAttribution = 'onAppOpenAttribution',
-  onAttributionFailure = 'onAttributionFailure',
-  CONVERSION_CALLBACK = 'conversion_callback',
-  OAOA_CALLBACK = 'oaoa_callback',
-  UDL_CALLBACK = 'udl_callback',
-}
-
 export enum AFPurchaseType {
   oneTimePurchase = 'one_time_purchase',
   subscription = 'subscription',